+++ packages/rockets-server-auth/src/domains/user/application/commands/handlers/signup-user.handler.ts
@@ -1,6 +1,6 @@
 import { CommandBus, QueryBus } from '@nestjs/cqrs';
 import { Injectable } from '@nestjs/common';
-import { TransactionScope } from '@concepta/nestjs-repository';
+import { TransactionScope } from '@concepta/rockets-core';
 
 import {
   CreateUserCommand,